+++ src/app/(popup)/else/page.tsx
@@ -1,0 +1,10 @@
+
+
+export default function Page() {
+	  return (
+	<div>
+	  <h1>popup else</h1>
+	  가나다
+	</div>
+  );
+}
